Created the Coding and Comments
Completed the assignment: Prompt user to enter file name and added the file as a default if the user leaves this blank. Open the file and iterate through the lines of the file, remove whitespace (right), skip lines that begin with ‘From:’ but not those that begin with ‘From’ (with no :), and add 1 to count each iteration. Print the final count.